[chore] UI 변경, home, gallery 바꾸기
