+++ app/src/main/java/org/sopt/sample/presentation/home/HomeFragment.kt
@@ -1,74 +1,33 @@
 package org.sopt.sample.presentation.home
 
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import org.sopt.sample.R
-import org.sopt.sample.adapter.RepoAdapter
-import org.sopt.sample.data.Repo
+import androidx.recyclerview.widget.GridLayoutManager
+import org.sopt.sample.adapter.GalleryAdapter
+import org.sopt.sample.data.remote.ServicePool
+import org.sopt.sample.data.remote.response.ResponseUser
 import org.sopt.sample.databinding.FragmentHomeBinding
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class HomeFragment : Fragment() {
+
     private var _binding: FragmentHomeBinding? = null
     private val binding: FragmentHomeBinding
         get() = requireNotNull(_binding)
 
-    private val mockRepoList = listOf<Repo>(
-        Repo(
-            image = 0,
-            name = "Yubeen's Repository",
-            author = "",
-            type = Repo.TEXT_TYPE
-        ),
-        Repo(
-            image = R.drawable.github,
-            name = "FILL-IN",
-            author = "Yubeen-park",
-            type = Repo.REPO_TYPE
-        ),
-        Repo(
-            image = R.drawable.github,
-            name = "FILL-IN",
-            author = "Yubeen-park",
-            type = Repo.REPO_TYPE
-        ),
-        Repo(
-            image = R.drawable.github,
-            name = "FILL-IN",
-            author = "Yubeen-park",
-            type = Repo.REPO_TYPE
-        ),
-        Repo(
-            image = R.drawable.github,
-            name = "FILL-IN",
-            author = "Yubeen-park",
-            type = Repo.REPO_TYPE
-        ),
-        Repo(
-            image = R.drawable.github,
-            name = "FILL-IN",
-            author = "Yubeen-park",
-            type = Repo.REPO_TYPE
-        ),
-        Repo(
-            image = R.drawable.github,
-            name = "FILL-IN",
-            author = "Yubeen-park",
-            type = Repo.REPO_TYPE
-        ),
-        Repo(
-            image = R.drawable.github,
-            name = "FILL-IN",
-            author = "Yubeen-park",
-            type = Repo.REPO_TYPE
-        )
-    )
+
+    private lateinit var gridManager: GridLayoutManager
+    private lateinit var galleryAdapter: GalleryAdapter
+    private val userListService = ServicePool.userListService
 
     override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
+        inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         _binding = FragmentHomeBinding.inflate(inflater, container, false)
@@ -77,14 +36,47 @@ class HomeFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val adapter = RepoAdapter(requireContext())
-        binding.rvRepos.adapter = adapter
-        adapter.setRepoList(mockRepoList)
+        initRecyclerView()
+        initUserInfo()
+    }
+
+    private fun initUserInfo() {
+        userListService.getUserList().enqueue(object : Callback<ResponseUser> {
+            override fun onResponse(
+                call: Call<ResponseUser>,
+                response: Response<ResponseUser>
+            ) {
+                if (response.isSuccessful) {
+                    response.body()?.let {
+                        galleryAdapter.setItems(it.data)
+                    }
+
+                }
+
+            }
+
+            override fun onFailure(call: Call<ResponseUser>, t: Throwable) {
+                Log.e("Gallery FAIL", "mes : " + t.message)
+            }
+        })
+    }
+
+    private fun initRecyclerView() {
+        gridManager = GridLayoutManager(requireContext(), SPAN_COUNT)
+        galleryAdapter = GalleryAdapter()
+        binding.rvGallery.apply {
+            layoutManager = gridManager
+            adapter = galleryAdapter
+        }
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
+    }
+
+    companion object {
+        const val SPAN_COUNT = 2
     }
 
 }
