+++ app/src/main/res/layout/activity_login.xml
@@ -20,7 +20,7 @@
             android:layout_marginTop="100dp"
             android:layout_width="wrap_content"
             android:text="@string/main_title"
-            android:textSize="20sp"
+            android:textSize="30sp"
             android:textStyle="bold"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
