+++ app/src/main/res/layout/activity_sign_up.xml
@@ -20,7 +20,7 @@
             android:id="@+id/txt_signup"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="40dp"
+            android:layout_marginTop="60dp"
             android:text="@string/signup_btn"
             android:textSize="20sp"
             android:textStyle="bold"
@@ -33,7 +33,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="40dp"
-            android:layout_marginTop="24dp"
+            android:layout_marginTop="40dp"
             android:text="@string/email"
             android:textStyle="bold"
             app:layout_constraintStart_toStartOf="parent"
