+++ app/src/main/res/layout/fragment_gallery.xml
@@ -7,8 +7,9 @@
     tools:context=".presentation.home.GalleryFragment">
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_gallery"
+        android:id="@+id/rv_repos"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         android:orientation="vertical" />
 </androidx.constraintlayout.widget.ConstraintLayout>
