+++ app/src/main/res/values/strings.xml
@@ -24,8 +24,7 @@
     <string name="id">ID</string>
     <string name="email">EMAIL</string>
 
-    <string name="signup_title">Welcome to SOPT</string>
-    <string name="main_title">Welcome to SOPT</string>
+    <string name="main_title">SOPT 31th</string>
     <string name="login">LOGIN</string>
     <string name="signup">SIGNUP</string>
     <string name="home">HOME</string>
@@ -38,6 +37,8 @@
     <string name="login_error">아이디, 비밀번호를 다시 확인해주세요!</string>
     <string name="signup_error">회원가입이 불가합니다</string>
     <string name="network_error">네트워크 에러</string>
+
+    <string name="home_title">이용자 목록</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
